[GUI] NBA Stat Tracker
This is the first major release of the NBA Stat Tracker. This application is intended for users to compare historical and/or live data on players. It supports a game log of the most five recent games, live game information, and a global player search.

Thanks to @swar from the public nba_api repo, the information that gets updated are from those endpoints. All issues can be reported and will be address.

Note, this is still a heavily under-development application. There are many things are still in the works but for now, this will do.